Clean-up, Tests & Readme
