+++ defluencer/src/signature_system/dag_jose.rs
@@ -125,83 +125,3 @@ impl TryFrom<RawJWS> for JsonWebSignature {
         })
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::signature_system::{
-        dag_jose::{JsonWebSignature, Signature},
-        IPNSSignature, SignatureSystem,
-    };
-
-    use super::*;
-
-    use ipfs_api::IpfsService;
-    use linked_data::signature::{AlgorithmType, CurveType, JsonWebKey, KeyType};
-    use rand_core::OsRng;
-
-    use ed25519_dalek::Keypair;
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 4)]
-    async fn ed25519_roundtrip() {
-        let mut csprng = OsRng::default();
-        let keypair: Keypair = Keypair::generate(&mut csprng);
-        let ori_pub_key = keypair.public.clone();
-
-        let ipfs = IpfsService::default();
-
-        let system = IPNSSignature::new(ipfs.clone(), keypair);
-
-        let cid =
-            Cid::try_from("bafybeig6xv5nwphfmvcnektpnojts33jqcuam7bmye2pb54adnrtccjlsu").unwrap();
-
-        let result = system.sign(cid).await.unwrap();
-
-        let raw: RawJWS = ipfs.dag_get(result, Option::<&str>::None).await.unwrap();
-
-        let signing_input = format!("{}.{}", raw.payload, raw.signatures[0].protected);
-
-        let JsonWebSignature {
-            payload,
-            mut signatures,
-            link: _,
-        } = raw.try_into().unwrap();
-
-        assert_eq!(payload, cid);
-
-        let Signature { header, signature } = signatures.remove(0);
-
-        let Header {
-            algorithm,
-            json_web_key,
-        } = header;
-
-        assert!(algorithm.unwrap() == AlgorithmType::EdDSA);
-
-        let JsonWebKey {
-            key_type,
-            curve,
-            x,
-            y: _,
-        } = json_web_key.unwrap();
-
-        assert!(key_type == KeyType::OctetString);
-        assert!(curve == CurveType::Ed25519);
-
-        let data = Base::Base64Url.decode(x).unwrap();
-
-        let public_key = ed25519_dalek::PublicKey::from_bytes(&data).unwrap();
-
-        assert_eq!(ori_pub_key, public_key);
-
-        let signature = ed25519_dalek::Signature::from_bytes(&signature).unwrap();
-
-        assert!(public_key
-            .verify(signing_input.as_bytes(), &signature)
-            .is_ok())
-    }
-
-    #[test]
-    fn k256_roundtrip() {
-        todo!()
-    }
-}
