+++ linked-data/Cargo.toml
@@ -6,13 +6,8 @@ edition = "2021"
 description = "Rust implementation of Defluencer linked data."
 
 [dependencies]
-either = { version = "1.6", default-features = false, features = ["serde"] }
+either = { version = "1", default-features = false, features = ["serde"] }
 cid = { version = "0.8", default-features = false, features = ["std"] }
-libsecp256k1 = { version = "0.7", default-features = false, features = ["lazy-static-context"] }
-serde = { version = "1.0", default-features = false, features = ["derive"] }
-serde_json = { version = "1.0" }
-serde_with = { version = "1.12", default-features = false, features = ["macros"] }
-tiny-keccak = { version = "2.0", default-features = false, features = ["keccak"] }
-
-[target.'cfg(target_arch = "wasm32-unknow-unknow")'.dependencies]
-getrandom = { version = "*", default-features = false, features = ["js"] }
+serde = { version = "1", default-features = false, features = ["derive"] }
+serde_json = { version = "1", default-features = false, features = [] }
+serde_with = { version = "1", default-features = false, features = ["macros"] }
