+++ linked-data/src/lib.rs
@@ -50,13 +50,3 @@ impl From<Cid> for IPLDLink {
         Self { link: cid }
     }
 }
-
-/// Compute the Keccak-256 hash of input bytes.
-pub fn keccak256(bytes: &[u8]) -> [u8; 32] {
-    use tiny_keccak::{Hasher, Keccak};
-    let mut output = [0u8; 32];
-    let mut hasher = Keccak::v256();
-    hasher.update(bytes);
-    hasher.finalize(&mut output);
-    output
-}
